abstract authentication
